+++ sqlalchemy_utils/__init__.py
@@ -24,7 +24,6 @@ from .functions import (
     get_referencing_foreign_keys,
     get_tables,
     group_foreign_keys,
-    has_any_changes,
     has_changes,
     has_index,
     has_unique_index,
@@ -80,7 +79,7 @@ from .types import (
 from .models import Timestamp
 
 
-__version__ = '0.27.7'
+__version__ = '0.27.8'
 
 
 __all__ = (
@@ -114,7 +113,6 @@ __all__ = (
     get_referencing_foreign_keys,
     get_tables,
     group_foreign_keys,
-    has_any_changes,
     has_changes,
     has_index,
     identity,