+++ CHANGES.rst
@@ -8,6 +8,7 @@ Here you can see the full list of changes between each SQLAlchemy-Utils release.
 ^^^^^^^^^^^^^^^^^^^
 
 - Added is_loaded utility function
+- Removed deprecated has_any_changes
 
 
 0.27.7 (2014-11-03)